requirements.txt
pandas>=2.0.3
tenacity>=8.2.3
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.3
flask[async]>=2.0.1
playwright>=1.42.0
hypercorn>=0.14.4
httpx>=0.24.1
anyio>=3.7.1
SQLAlchemy>=2.0.23
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
APScheduler>=3.10.4
Flask-Mail>=0.9.1
